Updating the readme to include archival
I think it's time for us to archive this project so we stop getting security alerts for this project since we do not intend to maintain it. I thought I'd put a little note first before actually marking this as archived. Shall we keep the issues and other PR open? (I already closed all of dependabot's ones)